Make a start on the merge refactor
The observation is that the original code drop of the sample app had
too many indirections from the parser's to the Vulkan API. This stems
from their original use-case, and makes sense there, but for the CTS's
use-case, it adds a lot of unnecessary complexity, making
determination of which spec semantics map to what tests *very*
difficult to trace and debug, when the parsing is wrong.

This needs to happen step by step, here the picture info flags are
populated directly in the parser, and some redundancies cleared
away. It is now also managing the field flag across the DPB correctly,
clearing up some tricky validation errors in interleaved cases.